move `check_symbol` to `checks.rs`
